requirements.txt
numpy==1.23.1
torch==1.12.0
torchvision==0.13.0
tensorboard==2.9.1
imageio==2.15.0
imageio-ffmpeg==0.4.7
matplotlib==3.5.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
tqdm==4.62.3
PyYAML==6.0
imgviz==1.5.0
opencv-python==4.1.2.30
fast-pytorch-kmeans==0.1.6
seaborn==0.11.2